fold download/cleanup tests into test_tasks. replace some stale tests with these
